# biblioteca-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autenticación",
        "description": "Controlador para registro, login y autenticación con JWT",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/prestamos": {
            "get": {
                "tags": [
                    "Préstamos"
                ],
                "summary": "Obtener la lista de préstamos",
                "operationId": "c103fb2b828a21ac7796cdaeac0c750e",
                "responses": {
                    "200": {
                        "description": "Lista de préstamos (según el rol del usuario)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Préstamos"
                ],
                "summary": "Crear un nuevo préstamo",
                "operationId": "1de0f63e287e73b15038ee8adde7408b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "libro_id",
                                    "fecha_prestamo",
                                    "fecha_devolucion"
                                ],
                                "properties": {
                                    "libro_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_prestamo": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-24"
                                    },
                                    "fecha_devolucion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Préstamo creado correctamente"
                    },
                    "400": {
                        "description": "Libro no disponible"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/prestamos/{id}": {
            "get": {
                "tags": [
                    "Préstamos"
                ],
                "summary": "Obtener un préstamo específico",
                "operationId": "1ee09d4a31f8dc0071584c5b4e6045a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del préstamo"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Préstamos"
                ],
                "summary": "Actualizar un préstamo",
                "operationId": "f7c1aeecf30fdcbccdaa4073f39f8f45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fecha_devolucion": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-01"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "entregado"
                                        ],
                                        "example": "entregado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Préstamo actualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Préstamos"
                ],
                "summary": "Eliminar un préstamo",
                "operationId": "006bce2e998854174a47bf55ec1bb649",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Préstamo eliminado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/libros/{id}/ultimo-prestamo": {
            "get": {
                "tags": [
                    "Préstamos"
                ],
                "summary": "Consultar el último préstamo pendiente de un libro",
                "operationId": "aef82eac218837fb1dc0b483166d0b64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Último préstamo encontrado"
                    },
                    "404": {
                        "description": "No hay préstamo pendiente para ese libro"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}